pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mlb-digest-cli"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "MLB-StatsAPI>=1.9.0",
    "anthropic>=0.40.0",
    "feedparser>=6.0",
    "click>=8.1",
    "tenacity>=8.2",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "markdown>=3.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-Markdown>=3.6",
]

[project.scripts]
mlb-digest = "mlb_digest.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = ["docs/plans/multi-team-files"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "W",    # pycodestyle warnings
    "T20",  # no print() in production code
    "G",    # logging format — no f-strings, use .exception()
    "ANN",  # type annotations on public functions
    "B",    # bugbear — mutable defaults, bare except, etc.
    "SIM",  # simplification — early returns, collapsed ifs
    "UP",   # modern Python 3.11+ idioms
    "PT",   # pytest conventions
    "RUF",  # ruff-specific quality checks
    "C4",   # unnecessary comprehension wrapping
    "PIE",  # no-op statements, unnecessary pass
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ANN",   # test functions don't need full annotations
    "T20",   # print() is fine in tests for debugging
]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["live: hits the real MLB Stats API (deselect with -m 'not live')"]
addopts = "-m 'not live'"

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
    "pre-commit>=4.5.1",
]
